src/requirements.txt
adagio==0.2.4
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.11.1
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blis==0.7.9
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.3.0
catalogue==2.0.8
catboost==1.2.2
certifi==2023.11.17
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.20
cloudpickle==2.2.1
cmdstanpy==1.2.0
colorama==0.4.6
confection==0.0.4
contourpy==1.1.1
coverage==7.2.3
cryptography==40.0.2
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.7
Cython==3.0.2
dask==2023.9.3
dataclasses-json==0.5.7
distlib==0.3.6
distributed==2023.9.3
dnspython==2.3.0
docker==6.0.1
duckdb==0.7.1
duckduckgo-search==2.8.5
dukpy==0.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
faiss-cpu==1.7.3
fastapi==0.95.1
filelock==3.12.2
firebase-admin==6.1.0
flake8==6.0.0
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.43.1
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.9.2
fugue==0.8.6
fugue-sql-antlr==0.1.7
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
gptcache==0.1.11
graphviz==0.20.1
grpcio==1.54.0
grpcio-status==1.54.0
gTTS==2.3.1
h11==0.14.0
hnswlib==0.7.0
holidays==0.34
httplib2==0.22.0
httptools==0.5.0
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.5
langchain==0.0.141
langcodes==3.3.0
lightgbm==4.1.0
lightning-utilities==0.9.0
llvmlite==0.41.0
localpip==1.0.7
locket==1.0.0
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.8.0
mccabe==0.7.0
mod-wsgi==4.9.4
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nfoursid==1.0.1
nltk==3.8.1
numba==0.58.0
numpy==1.24.2
openai==0.27.2
openapi-schema-pydantic==1.2.4
orjson==3.8.10
p5py==1.0.0
packaging==23.1
pandas==2.0.0
partd==1.4.1
pathy==0.10.1
patsy==0.5.3
pdfminer.six==20221105
pdfplumber==0.9.0
pep517==0.13.0
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.8.1
playsound==1.2.2
plotly==5.17.0
pmdarima==2.0.3
polars==0.19.7
posthog==2.5.0
preshed==3.0.8
prophet==1.1.5
proto-plus==1.22.2
protobuf==4.22.3
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
PyJWT==2.6.0
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-SystemConfiguration==10.0
pyod==1.1.0
PyPAC==0.16.4
pyparsing==3.0.9
pypdf==3.7.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==2.0.9.post0
pytz==2023.3
PyYAML==6.0
qpd==0.4.4
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
shap==0.43.0
six==1.16.0
slicer==0.0.7
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.47
sqlglot==18.11.6
srsly==2.4.6
stanio==0.3.0
starlette==0.26.1
statsforecast==1.6.0
statsmodels==0.14.0
stripe==5.4.0
stumpy==1.12.0
sympy==1.11.1
tbats==1.1.3
tblib==2.0.0
tenacity==8.2.2
tensorboardX==2.6.2.2
thinc==8.1.9
threadpoolctl==3.1.0
tiktoken==0.3.3
tldextract==3.6.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.0
torchmetrics==1.2.0
torchvision==0.15.1
tornado==6.3.3
tqdm==4.65.0
transformers==4.28.1
triad==0.9.1
tsfresh==0.20.1
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
u8darts==0.26.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.23.1
Wand==0.6.11
wasabi==1.1.1
watchfiles==0.19.0
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
xarray==2023.9.0
xgboost==2.0.0
yarl==1.8.2
zict==3.0.0
zipp==3.17.0
zstandard==0.20.0
